// src/tools/model/ProfileInfo.tsx
export interface ProfileInfo {
  empNo: string;
  userId: string;
  userName: string;
  residentNumber: string;
  gender: string;
  engName: string;
  family: string;
  armyStart: string;
  armyEnd: string;
  armyBranch: string;
  birthday: string;
  address: string;
  highschool: string;
  gradSchool: string;
  education: string;
  loc: string;
  emailAuth: string;
  tel: string;
  position: string;
  spot: string;
  task: string;
  techGrade: string;
  scienceTechCertify: string;
  locDetail: string;
  hiredate: string;
  leavedate: string;
  emergencyTel: string;
  relation: string;
  multiMajor1: string;
  multiMajor2: string;
}