// src/tools/model/AccessProps.tsx
export interface AccessProps {
  userName: string;
  empNo: string;
  enteranceTime: string;
  leaveTime: string;
  day: string;
  accessPlace: string;
  accessMode: string;
  accessType: string;
  accessStatus: string;
  spot: string;
  task: string;
  timeTable: string;
  accessNum: number;
  position: string;
}